+++ lib/addressing/model.rb
@@ -3,7 +3,7 @@
 module Addressing
   module Model
     def validates_address_format(
-      fields: [:country_code, :administrative_area, :locality, :dependent_locality, :postal_code, :sorting_code, :address_line1, :address_line2, :organization, :given_name, :additional_name, :family_name, :locale], field_overrides: nil, **options
+      fields: [:country_code, :administrative_area, :locality, :dependent_locality, :postal_code, :sorting_code, :address_line1, :address_line2, :organization, :given_name, :additional_name, :family_name, :locale], field_overrides: nil, verify_postal_code: true, **options
     )
       fields = Array(fields)
       field_overrides ||= FieldOverrides.new({})
@@ -48,7 +48,7 @@ module Addressing
           subdivisions = verify_subdivisions(address, address_format, field_overrides)
 
           # Validate postal code.
-          verify_postal_code(address.postal_code, subdivisions, address_format) if used_fields.include?(AddressField::POSTAL_CODE)
+          verify_postal_code(address.postal_code, subdivisions, address_format) if used_fields.include?(AddressField::POSTAL_CODE) && verify_postal_code
         end
 
         define_method :verify_subdivisions do |address, address_format, field_overrides|