+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to `addressing` will be documented in this file.
 
 ## UNRELEASED
 
-- Sync data with commerceguys repository (v2.2.0)
+- Sync data with commerceguys repository (v2.2.2)
 
 ## 0.7.0 (2024-01-30)